webserver_main/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
paho-mqtt